requirements.txt
Django
matplotlib
nltk
pandas
pymongo
pyspark==3.4.3
seaborn
kafka-python
scikit-learn
py4j==0.10.9.7
python-dotenv